Refactor folder structure for SuccessAction
