+++ lib/routes/lnurl/payment/success_action/success_action_data.dart
@@ -1,0 +1,6 @@
+class SuccessActionData {
+  final String message;
+  final String? url;
+
+  SuccessActionData(this.message, this.url);
+}
