+++ lib/routes/lnurl/payment/success_action/success_action_dialog.dart
@@ -1,6 +1,6 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:c_breez/l10n/build_context_localizations.dart';
-import 'package:c_breez/utils/lnurl.dart';
+import 'package:c_breez/routes/lnurl/payment/success_action/success_action_data.dart';
 import 'package:c_breez/widgets/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher_string.dart';
